renamed keypair.js to keybundle.js
