+++ lib/hc-dpki/index.js
@@ -1,4 +1,4 @@
-exports.KeyBundle = require('./keypair').KeyBundle
+exports.KeyBundle = require('./keybundle').KeyBundle
 exports.work = require('./work')
 
 const seed = require('./seed')
