+++ lib/hc-dpki/seed.js
@@ -2,7 +2,7 @@ const { AsyncClass } = require('../n3h-common')
 const mosodium = require('../mosodium')
 const bip39 = require('bip39')
 
-const { KeyBundle } = require('./keypair')
+const { KeyBundle } = require('./keybundle')
 const util = require('./util')
 
 /**
