+++ lib/n3h-mod-dht-fullsync/gossip.js
@@ -38,14 +38,12 @@ function decodeLocToHashMap (bufferList) {
 /**
  */
 function encodePeerAddressList (bufferList, list) {
-  //bufferList.push(list.map(e => encodePeerAddress(e)))
   bufferList.push(list.map(e => e))
 }
 
 /**
  */
 function decodePeerAddressList (bufferList) {
-  // return bufferList.shift().map(e => decodePeerAddress(e))
   return bufferList.shift().map(e => e)
 }
 
@@ -55,7 +53,6 @@ function encodePeerAddressToTsMap (bufferList, map) {
   const keylist = []
   const tslist = []
   for (let [p, ts] of iterMap(map)) {
-    // keylist.push(encodePeerAddress(p))
     keylist.push(p)
     tslist.push(ts)
   }
@@ -76,7 +73,6 @@ function decodePeerAddressToTsMap (bufferList) {
   const map = new Map()
   for (let i = 0; i < keylist.length; ++i) {
     map.set(
-      // decodePeerAddress(keylist[i]),
       keylist[i],
       tslist[i]
     )
@@ -91,7 +87,6 @@ function encodePeerMap (bufferList, map) {
   const keylist = []
   const datalist = []
   for (let [p, d] of iterMap(map)) {
-    // keylist.push(encodePeerAddress(p))
     keylist.push(p)
     datalist.push(Buffer.from(d))
   }
@@ -112,7 +107,6 @@ function decodePeerMap (bufferList) {
   const map = new Map()
   for (let i = 0; i < keylist.length; ++i) {
     map.set(
-      // decodePeerAddress(keylist[i]),
       keylist[i],
       datalist[i].toString()
     )
